Implement basic naive replacement algorithm
This attempt just reads through the file input buffer looking at raw
bytes. If a given byte matches the first byte of the replacement
token, it checks for a byte-for-byte replacement of the rest of the
replacement token by reading ahead within the buffer.

There's no provision for a match spanning a buffer read boundary
(though if the iterator is halfway through the read buffer when an
initial match happens, it DOES shift and refill the read buffer).

It also does not handle binary mismatches, e.g., if the replacement
token is in Unicode but the file codes the character according to a
given codepage, or if the replacement token is composed Unicode but
the file codes for the same character in decomposed Unicode.

Lastly, this attempt does not account for any RTF state (e.g., how
many bytes to skip in multiple-encoding mode, whether we're in a
paragraph or the middle of binary picture data, etc.