Refactor charsets, rip out token process
The previous iteration on this design tried to be too clever with
getting the next RTF text, command, or other token, making the code
unreadable after a break of a couple months. Since this is a part-time
project, I will never make progress if I can't read the code and
understand where I was and what I was doing. Starting that part over...

Additionally, experimenting with different ways to factor out support
for character sets/code pages. This code is verbose and hurts ability
to read the code if it's in the main processing module.