Implement parsing by state machine on the stack
This is a significant update. Rather than trying anything fancy with
state tracking or token parsing on the fly or oracle read-aheads, all
we do is this:
	- If we encounter a { or }, dispatch a helper function for
          scope changes.
        - If we encounter a \, dispatch a helper function for reading
          and then processing RTF commands (including commands to
          insert text).
        - For anything else, dispatch a helper function for reading
          text.

        - When we've done that, check to see if we have a complete
          match, have invalidated the potential for a match, or
          aren't sure yet.

This refactor has the benefit of using the call stack as a built-in
state machine for free. We don't need to keep track of what phase
of processing we're in manually if it's inherent in the context of
the function we're in.

We also add preliminary UTF-8 support and make some progress on
charset/code page support.

THIS VERSION COMPILES AND HANDLES MOST ENGLISH LANGUAGE FILES
CORRECTLY.  Still a lot of work to be done, but this is the most
progress I've made on this in years.